% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{file_paths}
\alias{file_paths}
\title{file_paths}
\usage{
file_paths(training_files, dat_path, station_list_path)
}
\arguments{
\item{training_files}{character list of the names of file types being used to provide training data}

\item{dat_path}{path to the root directory for all of the files being pulled for creation of the training model}

\item{station_list_path}{path to the station list that was created by EQTransformer}
}
\value{

}
\description{
file_paths
}
