% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formStead.R
\name{form_STEAD}
\alias{form_STEAD}
\title{form_STEAD}
\usage{
form_STEAD(STEAD_data)
}
\arguments{
\item{pre_hdf5_filtered}{output from joining_sac2STEAD}
}
\description{
Reformats attributes for HDF5 and csv format
}
