% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_data.R
\name{header_data}
\alias{header_data}
\title{header_data}
\usage{
header_data(sac_path, wrangled_SEISMOS)
}
\arguments{
\item{sac_path}{path to the directory where you plan to store the SAC files}

\item{wrangled_SEISMOS}{output from wrangle_SEISMOS}
}
\value{

}
\description{
header_data
}
