% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trainer_hdf5.R
\name{make_trainer_hdf5}
\alias{make_trainer_hdf5}
\title{make_trainer_hdf5}
\usage{
make_trainer_hdf5(h5_file_path, meta_data, signal)
}
\arguments{
\item{h5_file_path}{path, with new file name included, where h5 file to be created}

\item{meta_data}{output from form_STEAD}

\item{signal}{sac_EV column values from sac2STEAD}
}
\value{

}
\description{
make_trainer_hdf5
}
