% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac2STEAD.R
\name{sac2STEAD}
\alias{sac2STEAD}
\title{sac2STEAD}
\usage{
sac2STEAD(sac2preStead, wrangle_SEISMOS)
}
\arguments{
\item{sac2preStead}{Dataframe containing the data from the initial mseed files, converted to SAC, merged into 24 hr segments and data needed for merging with signal labels.}

\item{wrangle_SEISMOS}{tidy data from wrangle_SEISMOS}
}
\description{
Cuts, resamples, and properly names the SAC data and signal labels
}
