% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mseed2sac.R
\name{use_mseed2sac}
\alias{use_mseed2sac}
\title{use_mseed2sac}
\usage{
use_mseed2sac(mseed_path, sac_path, executable_path, wrangled_SEISMOS)
}
\arguments{
\item{mseed_path}{root path where the mseed files where downloaded from EQTransformer}

\item{sac_path}{path to the directory where you plan to store the SAC files}

\item{wrangled_SEISMOS}{output from wrangle_SEISMOS}

\item{exicutable_path}{path to the directory that contains the mseed2sac executable from IRIS}
}
\value{

}
\description{
use_mseed2sac
}
