% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_SEISMOS.R
\name{wrangle_SEISMOS}
\alias{wrangle_SEISMOS}
\title{wrangle_SEISMOS}
\usage{
wrangle_SEISMOS(file_paths)
}
\arguments{
\item{file_paths}{list of paths to the dat files used for training}
}
\value{

}
\description{
wrangle_SEISMOS
}
